# Optimizing the sliding window in the frontend.